These files are for figuring out why the stdout from the subprocess call are buffered
The new FT_test2.ui is nicer and I'll use it to try a pydm_label rather than QT5 label when I get it working.
Testy2.py is a routine that prints out every few seconds to stdout for testing.
SubPCallTest.py and CommMicroNoCA.py are pydm code which tests the changes in a pydm environment.  Buffering seems to be a problem.
Test3.py tests for line by line stdout in a python environment.